Pluggable indexers (#16)
* top level index method

* sending to indexers

* makes db connection an option

* starting to build actual indexers

* pluggable indexers
* bigdecimal default instead of from_str

* updated db type

* renamed root registry index method

* renamed Index to IndexMessage

* commented indexer trait

* made events non-optional and a custom type

* logging

* uses registry instead of passing the PgConnecton around

* registry key refactoring

+++ src/db/schema.rs
@@ -39,7 +39,7 @@ table! {
         id -> Int4,
         address -> Text,
         token -> Text,
-        balance -> Int8,
+        balance -> Numeric,
     }
 }
